Added changes from review comments
