+++ build.sbt
@@ -14,15 +14,15 @@ val akkaPersistenceInMemVersion     = "2.5.1.1"
 val akkaPersistenceCassandraVersion = "0.55"
 val akkaHttpVersion                 = "10.0.10"
 val akkaHttpCirceVersion            = "1.18.0"
-val elasticSearchVersion            = "5.5.3"
-val log4jVersion                    = "2.8.2"
+val elasticSearchVersion            = "6.1.1"
+val log4jVersion                    = "2.10.0"
+val commonsio                       = "1.3.2"
 
 lazy val catsCore           = "org.typelevel"                   %% "cats-core"                           % catsVersion
 lazy val circeCore          = "io.circe"                        %% "circe-core"                          % circeVersion
 lazy val circeParser        = "io.circe"                        %% "circe-parser"                        % circeVersion
 lazy val circeGenericExtras = "io.circe"                        %% "circe-generic-extras"                % circeVersion
 lazy val circeJava8         = "io.circe"                        %% "circe-java8"                         % circeVersion
-lazy val circeOptics        = "io.circe"                        %% "circe-optics"                        % circeVersion
 lazy val scalaTest          = "org.scalatest"                   %% "scalatest"                           % scalaTestVersion
 lazy val shapeless          = "com.chuusai"                     %% "shapeless"                           % shapelessVersion
 lazy val monixEval          = "io.monix"                        %% "monix-eval"                          % monixVersion
@@ -154,24 +154,23 @@ lazy val iam = project
                                 scalaTest   % Test)
   )
 
-lazy val indexTypes = project
-  .in(file("modules/index/types"))
+lazy val queryTypes = project
+  .in(file("modules/query-types"))
   .settings(
-    name                := "commons-index-types",
-    moduleName          := "commons-index-types",
-    libraryDependencies ++= Seq(akkaHttp, catsCore, circeCore, scalaTest % Test, circeGenericExtras % Test)
+    name                := "commons-query-types",
+    moduleName          := "commons-query-types",
+    libraryDependencies ++= Seq(catsCore, circeCore, shapeless, scalaTest % Test, circeGenericExtras % Test)
   )
 
 lazy val elasticClient = project
-  .in(file("modules/index/elastic-client"))
-  .dependsOn(http, indexTypes, test)
+  .in(file("modules/elastic-client"))
+  .dependsOn(http, queryTypes, test)
   .settings(
     name       := "elastic-client",
     moduleName := "elastic-client",
     libraryDependencies ++= Seq(
       akkaStream,
       circeCore,
-      circeOptics,
       circeParser                % Test,
       circeGenericExtras         % Test,
       akkaTestKit                % Test,
@@ -179,15 +178,15 @@ lazy val elasticClient = project
       "org.apache.logging.log4j" % "log4j-api" % log4jVersion % Test,
       "org.apache.logging.log4j" % "log4j-core" % log4jVersion % Test,
       "org.elasticsearch"        % "elasticsearch" % elasticSearchVersion % Test,
-      "org.elasticsearch.client" % "rest" % elasticSearchVersion % Test,
-      "org.elasticsearch.plugin" % "transport-netty3-client" % elasticSearchVersion % Test,
-      "org.apache.commons"       % "commons-io" % "1.3.2" % Test
+      "org.elasticsearch.client" % "elasticsearch-rest-client" % elasticSearchVersion % Test,
+      "org.elasticsearch.plugin" % "transport-netty4-client" % elasticSearchVersion % Test,
+      "org.apache.commons"       % "commons-io" % commonsio % Test
     )
   )
 
 lazy val sparqlClient = project
-  .in(file("modules/index/sparql-client"))
-  .dependsOn(http, indexTypes)
+  .in(file("modules/sparql-client"))
+  .dependsOn(http, queryTypes)
   .settings(
     name       := "sparql-client",
     moduleName := "sparql-client",
@@ -235,7 +234,7 @@ lazy val root = project
              http,
              test,
              service,
-             indexTypes,
+             queryTypes,
              elasticClient,
              sparqlClient,
              shaclValidator,
