+++ modules/elastic-client/src/main/scala/ch/epfl/bluebrain/nexus/commons/es/client/ElasticBaseClient.scala
