+++ modules/elastic-client/src/test/resources/index_payload.json
@@ -1,0 +1,21 @@
+{
+  "mappings": {
+    "doc" : {
+      "properties": {
+        "key": {
+          "type": "keyword"
+        },
+        "/key2": {
+          "type": "keyword"
+        },
+        "age": {
+          "type": "integer"
+        }
+      }
+    }
+  },
+  "settings": {
+    "number_of_shards": "5",
+    "number_of_replicas": "1"
+  }
+}
