+++ modules/elastic-client/src/test/resources/simple_query.json
@@ -1,0 +1,13 @@
+{
+  "query" : {
+    "constant_score" : {
+      "filter" : {
+        "bool" : {
+          "must" : {
+            "term" : {"{{k}}" : "{{v}}"}
+          }
+        }
+      }
+    }
+  }
+}
