+++ modules/elastic-client/src/test/scala/ch/epfl/bluebrain/nexus/commons/es/client/ElasticDecoderSpec.scala
