--- modules/index/elastic-client/src/test/resources/index_payload.json
@@ -1,25 +1,0 @@
-{
-  "mappings": {
-    "aa": {},
-    "bb": {
-      "_parent": {
-        "type": "aa"
-      },
-      "properties": {
-        "http://localhost/voc/key": {
-          "type": "keyword"
-        },
-        "http://localhost/voc/key2": {
-          "type": "keyword"
-        },
-        "age": {
-          "type": "integer"
-        }
-      }
-    }
-  },
-  "settings": {
-    "number_of_shards": "5",
-    "number_of_replicas": "1"
-  }
-}
