--- modules/index/elastic-client/src/test/resources/query.json
@@ -1,16 +1,0 @@
-{
-  "query" : {
-    "constant_score" : {
-      "filter" : {
-        "bool" : {
-          "must" : {
-            "term" : {"http://localhost/voc/{{key1}}" : "{{value1}}"}
-          },
-          "must" : {
-            "term" : {"http://localhost/voc/{{key2}}" : "{{value2}}"}
-          }
-        }
-      }
-    }
-  }
-}
