--- modules/index/elastic-client/src/test/resources/simple_query.json
@@ -1,13 +1,0 @@
-{
-  "query" : {
-    "constant_score" : {
-      "filter" : {
-        "bool" : {
-          "must" : {
-            "term" : {"http://localhost/voc/{{k}}" : "{{v}}"}
-          }
-        }
-      }
-    }
-  }
-}
