+++ modules/query-types/src/main/scala/ch/epfl/bluebrain/nexus/commons/types/search/Pagination.scala
