+++ modules/query-types/src/test/scala/ch/epfl/bluebrain/nexus/commons/types/search/QueryResultSpec.scala
