+++ modules/sparql-client/src/main/scala/ch/epfl/bluebrain/nexus/commons/sparql/client/SparqlFailure.scala
