+++ modules/sparql-client/src/test/resources/RWStore.properties
