+++ modules/sparql-client/src/test/resources/application.conf
