+++ modules/sparql-client/src/test/resources/index.properties
