+++ modules/sparql-client/src/test/resources/ld.json
