+++ modules/sparql-client/src/test/resources/logback.xml
