+++ modules/sparql-client/src/test/scala/ch/epfl/bluebrain/nexus/commons/sparql/client/SparqlClientSpec.scala
